packages/mcp-server-hub/pyproject.toml
[project]
name = "mcp-server-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.27.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.18.0",
    "tavily-python>=0.7.24",
    "trafilatura>=2.0.0",
]
